crm/pyproject.toml
[tool.poetry]
name = "vovacrm"
version = "0.1.0"
description = ""
authors = ["hikmatillo <dolgovdaniil007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.1.0"}
django-import-export = "^3.2.0"
django-jazzmin = "^2.6.0"
babel = "^2.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
